Move top one/multi syllable lists into rules/
