+++ src/lib/rules/consts/topMultiSyllableTranslations.ts
