+++ src/lib/rules/consts/topOneSyllableTranslations.ts
